+++ NEWS.md
@@ -11,6 +11,7 @@
 
 - post-customization of color and fill no longer shows warnings like "Scale for 'fill' is already present. Adding another scale for 'fill', which will replace the existing scale" ([@MarcinKosinski, #11](https://github.com/kassambara/survminer/issues/11)).
 - now, post-customization of survival curve colors will automatically affect the risk table y axis text colors ([@MarcinKosinski, #11](https://github.com/kassambara/survminer/issues/12)).
+- Default value for the argument `risk.table.y.text.col` is now TRUE.
    
    
 ## Bug fixes