+++ NEWS.md
@@ -6,6 +6,8 @@
 
 - New `ggcoxfunctional` function that displays graphs of continuous explanatory variable against martingale residuals of null
  cox proportional hazards model, for each term in of the right side of input formula. This might help to properly choose the functional form of continuous variable in cox model, since fitted lines with `lowess` function should be linear to satisfy cox proportional hazards model assumptions. ([@MarcinKosinski, #14](https://github.com/kassambara/survminer/issues/14))
+ 
+- New function `theme_classic2`: ggplot2 classic theme with axis line. This function replaces ggplot2::theme_classic, which does no longer display axis lines (since ggplot2 v2.1.0)
    
 ## Minor changes