+++ NEWS.md
@@ -1,3 +1,15 @@
+# survminer 0.2.1.900
+    
+    
+## New features
+   
+   
+## Minor changes
+   
+   
+## Bug fixes
+    
+    
 # survminer 0.2.1
 
 ## New features